Web dash/requirements.txt
fastapi
uvicorn
gunicorn
requests
python-dotenv
numpy
tensorflow
jinja2
python-multipart
opencv-python
langchain-groq
langchain
pytz